+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Tests/MessageHub/Models/Shared/CimValidationErrorCodeFactoryTests.cs
@@ -48,7 +48,7 @@ namespace GreenEnergyHub.Charges.Tests.MessageHub.Models.Shared
         [InlineAutoMoqData(ValidationRuleIdentifier.CommandSenderMustBeAnExistingMarketParticipant, ReasonCode.D02)]
         [InlineAutoMoqData(ValidationRuleIdentifier.DocumentTypeMustBeRequestChangeOfPriceList, ReasonCode.D02)]
         [InlineAutoMoqData(ValidationRuleIdentifier.MeteringPointDoesNotExist, ReasonCode.E10)]
-        [InlineAutoMoqData(ValidationRuleIdentifier.ChargeDoesNotExist, ReasonCode.E0I)]
+        [InlineAutoMoqData(ValidationRuleIdentifier.ChargeDoesNotExistOnLinkStartDate, ReasonCode.E0I)]
         [InlineAutoMoqData(ValidationRuleIdentifier.ChargeLinkUpdateNotYetSupported, ReasonCode.D13)]
         [InlineAutoMoqData(ValidationRuleIdentifier.SubsequentBundleOperationsFail, ReasonCode.D14)]
         [InlineAutoMoqData(ValidationRuleIdentifier.TransparentInvoicingIsNotAllowedForFee, ReasonCode.D67)]
@@ -64,8 +64,9 @@ namespace GreenEnergyHub.Charges.Tests.MessageHub.Models.Shared
         [InlineAutoMoqData(ValidationRuleIdentifier.TaxIndicatorMustBeFalseForFee, ReasonCode.D14)]
         [InlineAutoMoqData(ValidationRuleIdentifier.TaxIndicatorMustBeFalseForSubscription, ReasonCode.D14)]
         [InlineAutoMoqData(ValidationRuleIdentifier.ChargeTypeTariffTaxIndicatorOnlyAllowedBySystemOperator, ReasonCode.E0I)]
-        [InlineAutoMoqData(ValidationRuleIdentifier.UpdateTaxTariffOnlyAllowedBySystemOperator, ReasonCode.E0I)]
         [InlineAutoMoqData(ValidationRuleIdentifier.PriceSeriesResolutionMustMatchChargeResolution, ReasonCode.D14)]
+        [InlineAutoMoqData(ValidationRuleIdentifier.PointsRequired, ReasonCode.E87)]
+        [InlineAutoMoqData(ValidationRuleIdentifier.ChargeDoesNotExist, ReasonCode.D14)]
         public void Create_ReturnsExpectedCode(
             ValidationRuleIdentifier identifier,
             ReasonCode expected,