+++ source/Energinet.DataHub.Charges.Libraries/source/Energinet.DataHub.Charges.Clients.Registration/documents/release-notes/release-notes.md
@@ -1,5 +1,10 @@
 # Energinet.DataHub.Charges.Clients.Registrations Release notes
 
+## Version 6.0.0
+
+`SearchChargesAsync` now returns one `ChargeV1Dto` per period per charge.
+Removed `GetChargesAsync()` from `ChargesClient`
+
 ## Version 5.0.1
 
 Added `Role` to `MarketParticipantV1Dto`